+++ storefront/src/main/java/com/shopflix/storefront/forms/AddToCartForm.java
@@ -11,7 +11,7 @@ public class AddToCartForm
 
     @NotNull(message = "{basket.error.quantity.notNull}")
     @Min(value = 1, message = "{basket.error.quantity.invalid}")
-    private long qty;
+    private int qty;
 
     public Long getProductId()
     {
@@ -33,12 +33,12 @@ public class AddToCartForm
         this.variantId = variantId;
     }
 
-    public long getQty()
+    public int getQty()
     {
         return qty;
     }
 
-    public void setQty(long qty)
+    public void setQty(int qty)
     {
         this.qty = qty;
     }
